full lib

+++ libft/src/ft_memset.c
@@ -6,17 +6,17 @@
 /*   By: yscheupl <yscheupl@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/07 16:05:51 by yscheupl          #+#    #+#             */
-/*   Updated: 2025/05/07 16:51:12 by yscheupl         ###   ########.fr       */
+/*   Updated: 2025/06/05 14:43:48 by yscheupl         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include"../libft.h"
+#include "../libft.h"
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int	i;
-	char * tmp;
-	
+	size_t	i;
+	char	*tmp;
+
 	i = 0;
 	tmp = (char *) s;
 	while (i < n)
